add testcase
